MOS-28489 - Integrated review comments
Former-commit-id: c56e103601e462898d5bcfd6c7089b0560bbd7ec
Former-commit-id: 2621ff5d65e8b0d02839a595221ab97c18110cb6

+++ authentication/authentication-common/src/main/java/io/mosip/authentication/common/service/repository/UinHashSaltRepo.java
@@ -22,5 +22,5 @@ public interface UinHashSaltRepo extends JpaRepository<UinHashSalt, Integer> {
 	 * @return String salt
 	 */
 	@Query("select salt from UinHashSalt where id = :id")
-	public String retrieveSaltById(@Param("id") int id);
+	public String retrieveSaltById(@Param("id") long id);
 }
